Tests/LibWeb/Text/expected/IntersectionObserver/box-without-layout-node-should-not-intersect.txt

flip #container to visible
#nested is visible
